allow annotate models to be swapped with custom implementations

+++ patterns/annotate/migrations/0001_initial.py
@@ -3,45 +3,53 @@ import datetime
 from south.db import db
 from south.v2 import SchemaMigration
 from django.db import models
+from wq.db.patterns.base import swapper
 
 class Migration(SchemaMigration):
 
     def forwards(self, orm):
-        
-        # Adding model 'AnnotationType'
-        db.create_table('wq_annotationtype', (
-            ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
-            ('name', self.gf('django.db.models.fields.CharField')(max_length=255)),
-            ('contenttype', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['contenttypes.ContentType'], null=True, blank=True)),
-        ))
-        db.send_create_signal('annotate', ['AnnotationType'])
 
-        # Adding model 'Annotation'
-        db.create_table('wq_annotation', (
-            ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
-            ('type', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['annotate.AnnotationType'])),
-            ('value', self.gf('django.db.models.fields.TextField')(null=True, blank=True)),
-            ('content_type', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['contenttypes.ContentType'])),
-            ('object_id', self.gf('django.db.models.fields.PositiveIntegerField')()),
-        ))
-        db.send_create_signal('annotate', ['Annotation'])
+        if not swapper.is_swapped('annotate', 'AnnotationType'):
 
+            # Adding model 'AnnotationType'
+            db.create_table('wq_annotationtype', (
+                ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+                ('name', self.gf('django.db.models.fields.CharField')(max_length=255)),
+                ('contenttype', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['contenttypes.ContentType'], null=True, blank=True)),
+            ))
+            db.send_create_signal('annotate', ['AnnotationType'])
 
-        # Create index
-        db.execute("""CREATE INDEX wq_annotation_idx ON wq_annotation
-                         (content_type_id, object_id)""")
+        if not swapper.is_swapped('annotate', 'Annotation'):
+            AnnotationType = swapper.load_model('annotate', 'AnnotationType', orm)
+
+            # Adding model 'Annotation'
+            db.create_table('wq_annotation', (
+                ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+
+                ('type', self.gf('django.db.models.fields.related.ForeignKey')(to=AnnotationType)),
+                ('value', self.gf('django.db.models.fields.TextField')(null=True, blank=True)),
+                ('content_type', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['contenttypes.ContentType'])),
+                ('object_id', self.gf('django.db.models.fields.PositiveIntegerField')()),
+            ))
+            db.send_create_signal('annotate', ['Annotation'])
+
+            # Create index
+            db.execute("""CREATE INDEX wq_annotation_idx ON wq_annotation
+                             (content_type_id, object_id)""")
 
 
     def backwards(self, orm):
 
-        # Drop index
-        db.execute("DROP INDEX wq_annotation_idx;");
-        
-        # Deleting model 'AnnotationType'
-        db.delete_table('wq_annotationtype')
+        if not swapper.is_swapped('annotate', 'AnnotationType'):
+            # Deleting model 'AnnotationType'
+            db.delete_table('wq_annotationtype')
 
-        # Deleting model 'Annotation'
-        db.delete_table('wq_annotation')
+        if not swapper.is_swapped('annotate', 'Annotation'):
+            # Drop index
+            db.execute("DROP INDEX wq_annotation_idx;");
+        
+            # Deleting model 'Annotation'
+            db.delete_table('wq_annotation')
 
 
     models = {
